Refactor .gitconfig symlink+include structure
Previously, most dotfiles-provided git config values were found within
the main .gitconfig symlink, and local values were expected to be
added into the .gitconfig.local symlink.

Now, the dotfiles-provided git config is found within a symlink for
.gitconfig.dotfiles, and this file is included from within the main
.gitconfig symlink.

Although conventional wisdom is to append .local to local and/or
uncommitted files, this is problematic as the git-config subcommand
will attempt to write directly to the .gitconfig file in any case.
This refactoring will allow the git-config subcommand to behave
properly without altering the version-controlled dotfiles configs.